refactor: extract shared Mocha runner, clean up e2e test, add docs
- Extract createMochaRunner to deduplicate suite/index.ts files
- Lift phpFolder/tsFolder to suite-level vars in issue-381 test
- Map URI strings once instead of calling toString() twice
- Add multi-workspace config warning to README
