+++ packages/extension/src/TestDiscovery/TestFileDiscovery.ts
@@ -67,8 +67,8 @@ export class TestFileDiscovery {
         const folderUri = this.workspaceFolder.uri.toString();
         const newFiles = new Set(
             (await workspace.findFiles(pattern, exclude))
-                .filter((f) => f.toString().startsWith(folderUri))
-                .map((f) => f.toString()),
+                .map((f) => f.toString())
+                .filter((uri) => uri.startsWith(folderUri)),
         );
 
         // Remove files that no longer exist
